updated to have a modal component
